pyproject.toml
[project]
name = "anti-chaos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "aiosqlite==0.19.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-telegram-bot==20.7",
    "apscheduler==3.10.4",
    "cryptography==41.0.7",
    "greenlet>=3.3.0",
]
